+++ tests/integ/test_tuner.py
@@ -99,6 +99,8 @@ def _tune_and_deploy(
     warm_start_config=None,
     early_stopping_type="Off",
     instance_configs=None,
+    autotune=False,
+    hyperparameters_to_keep_static=None,
 ):
     tuner = _tune(
         kmeans_estimator,
@@ -108,6 +110,8 @@ def _tune_and_deploy(
         job_name=job_name,
         early_stopping_type=early_stopping_type,
         instance_configs=instance_configs,
+        autotune=autotune,
+        hyperparameters_to_keep_static=hyperparameters_to_keep_static,
     )
     _deploy(kmeans_train_set, sagemaker_session, tuner, early_stopping_type, cpu_instance_type)
 
@@ -138,6 +142,8 @@ def _tune(
     max_parallel_jobs=2,
     early_stopping_type="Off",
     instance_configs=None,
+    autotune=False,
+    hyperparameters_to_keep_static=None,
 ):
     with timeout(minutes=TUNING_DEFAULT_TIMEOUT_MINUTES):
 
@@ -151,6 +157,8 @@ def _tune(
                 max_parallel_jobs=max_parallel_jobs,
                 warm_start_config=warm_start_config,
                 early_stopping_type=early_stopping_type,
+                autotune=autotune,
+                hyperparameters_to_keep_static=hyperparameters_to_keep_static,
             )
         tuner.override_resource_config(instance_configs=instance_configs)
         records = kmeans_estimator.record_set(kmeans_train_set[0][:100])
@@ -174,6 +182,26 @@ def test_tuning_kmeans(
         cpu_instance_type,
         hyperparameter_ranges=hyperparameter_ranges,
         job_name=job_name,
+    )
+
+
+def test_tuning_kmeans_autotune(
+    sagemaker_session, kmeans_train_set, kmeans_estimator, hyperparameter_ranges, cpu_instance_type
+):
+    job_name = unique_name_from_base("test-tune-kmeans-autotune", 32)
+    _tune_and_deploy(
+        kmeans_estimator,
+        kmeans_train_set,
+        sagemaker_session,
+        cpu_instance_type,
+        hyperparameter_ranges=hyperparameter_ranges,
+        job_name=job_name,
+        autotune=True,
+        hyperparameters_to_keep_static=[
+            "local_lloyd_init_method",
+            "local_lloyd_num_trials",
+            "local_lloyd_tol",
+        ],
     )
 
 
@@ -590,6 +618,62 @@ def test_tuning_mxnet(
         )
 
         tuning_job_name = unique_name_from_base("tune-mxnet", max_length=32)
+        print("Started hyperparameter tuning job with name:" + tuning_job_name)
+        tuner.fit({"train": train_input, "test": test_input}, job_name=tuning_job_name)
+
+    best_training_job = tuner.best_training_job()
+    with timeout_and_delete_endpoint_by_name(best_training_job, sagemaker_session):
+        predictor = tuner.deploy(1, cpu_instance_type)
+        data = np.zeros(shape=(1, 1, 28, 28))
+        predictor.predict(data)
+
+
+@pytest.mark.slow_test
+def test_tuning_mxnet_autotune(
+    sagemaker_session,
+    mxnet_training_latest_version,
+    mxnet_training_latest_py_version,
+    cpu_instance_type,
+):
+    with timeout(minutes=TUNING_DEFAULT_TIMEOUT_MINUTES):
+        script_path = os.path.join(DATA_DIR, "mxnet_mnist", "mnist.py")
+        data_path = os.path.join(DATA_DIR, "mxnet_mnist")
+
+        estimator = MXNet(
+            entry_point=script_path,
+            role="SageMakerRole",
+            py_version=mxnet_training_latest_py_version,
+            instance_count=1,
+            instance_type=cpu_instance_type,
+            framework_version=mxnet_training_latest_version,
+            sagemaker_session=sagemaker_session,
+        )
+        estimator.set_hyperparameters(learning_rate=0.1)
+
+        hyperparameter_ranges = {}
+        objective_metric_name = "Validation-accuracy"
+        metric_definitions = [
+            {"Name": "Validation-accuracy", "Regex": "Validation-accuracy=([0-9\\.]+)"}
+        ]
+        tuner = HyperparameterTuner(
+            estimator,
+            objective_metric_name,
+            hyperparameter_ranges,
+            metric_definitions,
+            max_jobs=4,
+            max_parallel_jobs=2,
+            autotune=True,
+            hyperparameters_to_keep_static=None,
+        )
+
+        train_input = estimator.sagemaker_session.upload_data(
+            path=os.path.join(data_path, "train"), key_prefix="integ-test-data/mxnet_mnist/train"
+        )
+        test_input = estimator.sagemaker_session.upload_data(
+            path=os.path.join(data_path, "test"), key_prefix="integ-test-data/mxnet_mnist/test"
+        )
+
+        tuning_job_name = unique_name_from_base("tune-mxnet-autotune", max_length=32)
         print("Started hyperparameter tuning job with name:" + tuning_job_name)
         tuner.fit({"train": train_input, "test": test_input}, job_name=tuning_job_name)