+++ tests/unit/sagemaker/image_uris/expected_uris.py
@@ -80,6 +80,17 @@ def djl_framework_uri(repo, account, djl_version, primary_framework, region=REGI
     return IMAGE_URI_FORMAT.format(account, region, domain, repo, tag)
 
 
+def huggingface_llm_framework_uri(
+    repo,
+    account,
+    version,
+    tag,
+    region=REGION,
+):
+    domain = ALTERNATE_DOMAINS.get(region, DOMAIN)
+    return IMAGE_URI_FORMAT.format(account, region, domain, repo, tag)
+
+
 def base_python_uri(repo, account, region=REGION):
     domain = ALTERNATE_DOMAINS.get(region, DOMAIN)
     tag = "1.0"