+++ tests/unit/test_session.py
@@ -1718,6 +1718,104 @@ def test_train_with_sagemaker_config_injection(sagemaker_session):
     }
 
 
+def test_train_with_sagemaker_config_injection_no_kms_support(sagemaker_session):
+    """Tests that when the instance type for train does not support KMS, no KMS key is used."""
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_TRAINING_JOB
+
+    in_config = [
+        {
+            "ChannelName": "training",
+            "DataSource": {
+                "S3DataSource": {
+                    "S3DataDistributionType": "FullyReplicated",
+                    "S3DataType": "S3Prefix",
+                    "S3Uri": S3_INPUT_URI,
+                }
+            },
+        }
+    ]
+
+    out_config = {"S3OutputPath": S3_OUTPUT}
+
+    resource_config = {
+        "InstanceCount": INSTANCE_COUNT,
+        "InstanceType": "ml.g5.12xlarge",
+    }
+
+    stop_cond = {"MaxRuntimeInSeconds": MAX_TIME}
+    RETRY_STRATEGY = {"MaximumRetryAttempts": 2}
+    hyperparameters = {"foo": "bar"}
+    TRAINING_IMAGE_CONFIG = {
+        "TrainingRepositoryAccessMode": "Vpc",
+        "TrainingRepositoryAuthConfig": {
+            "TrainingRepositoryCredentialsProviderArn": "arn:aws:lambda:us-west-2:1234567897:function:test"
+        },
+    }
+
+    sagemaker_session.train(
+        image_uri=IMAGE,
+        input_mode="File",
+        input_config=in_config,
+        job_name=JOB_NAME,
+        output_config=out_config,
+        resource_config=resource_config,
+        hyperparameters=hyperparameters,
+        stop_condition=stop_cond,
+        metric_definitions=METRIC_DEFINITONS,
+        use_spot_instances=True,
+        checkpoint_s3_uri="s3://mybucket/checkpoints/",
+        checkpoint_local_path="/tmp/checkpoints",
+        enable_sagemaker_metrics=True,
+        environment=ENV_INPUT,
+        retry_strategy=RETRY_STRATEGY,
+        training_image_config=TRAINING_IMAGE_CONFIG,
+    )
+
+    _, _, actual_train_args = sagemaker_session.sagemaker_client.method_calls[0]
+
+    expected_role_arn = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"]["TrainingJob"]["RoleArn"]
+    expected_kms_key_id = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"]["TrainingJob"][
+        "OutputDataConfig"
+    ]["KmsKeyId"]
+    expected_vpc_config = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"]["TrainingJob"]["VpcConfig"]
+    expected_enable_network_isolation = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"]["TrainingJob"][
+        "EnableNetworkIsolation"
+    ]
+    expected_enable_inter_container_traffic_encryption = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"][
+        "TrainingJob"
+    ]["EnableInterContainerTrafficEncryption"]
+    expected_tags = SAGEMAKER_CONFIG_TRAINING_JOB["SageMaker"]["TrainingJob"]["Tags"]
+
+    assert actual_train_args["VpcConfig"] == expected_vpc_config
+    assert actual_train_args["HyperParameters"] == hyperparameters
+    assert actual_train_args["Tags"] == expected_tags
+    assert actual_train_args["AlgorithmSpecification"]["MetricDefinitions"] == METRIC_DEFINITONS
+    assert actual_train_args["AlgorithmSpecification"]["EnableSageMakerMetricsTimeSeries"] is True
+    assert (
+        actual_train_args["EnableInterContainerTrafficEncryption"]
+        == expected_enable_inter_container_traffic_encryption
+    )
+    assert actual_train_args["EnableNetworkIsolation"] == expected_enable_network_isolation
+    assert actual_train_args["EnableManagedSpotTraining"] is True
+    assert actual_train_args["CheckpointConfig"]["S3Uri"] == "s3://mybucket/checkpoints/"
+    assert actual_train_args["CheckpointConfig"]["LocalPath"] == "/tmp/checkpoints"
+    assert actual_train_args["Environment"] == ENV_INPUT
+    assert actual_train_args["RetryStrategy"] == RETRY_STRATEGY
+    assert (
+        actual_train_args["AlgorithmSpecification"]["TrainingImageConfig"] == TRAINING_IMAGE_CONFIG
+    )
+    assert actual_train_args["RoleArn"] == expected_role_arn
+    assert actual_train_args["ResourceConfig"] == {
+        "InstanceCount": INSTANCE_COUNT,
+        "InstanceType": "ml.g5.12xlarge",
+    }
+    assert actual_train_args["OutputDataConfig"] == {
+        "S3OutputPath": S3_OUTPUT,
+        "KmsKeyId": expected_kms_key_id,
+    }
+
+
 def test_train_pack_to_request_with_optional_params(sagemaker_session):
     in_config = [
         {
@@ -2646,7 +2744,7 @@ def test_create_endpoint_config_with_sagemaker_config_injection(sagemaker_sessio
         "endpoint-test",
         "simple-model",
         1,
-        "local",
+        "ml.p2.xlarge",
         data_capture_config_dict=data_capture_config_dict,
     )
     expected_data_capture_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
@@ -2665,7 +2763,7 @@ def test_create_endpoint_config_with_sagemaker_config_injection(sagemaker_sessio
                 "VariantName": "AllTraffic",
                 "InitialVariantWeight": 1,
                 "InitialInstanceCount": 1,
-                "InstanceType": "local",
+                "InstanceType": "ml.p2.xlarge",
             }
         ],
         DataCaptureConfig={
@@ -2673,6 +2771,46 @@ def test_create_endpoint_config_with_sagemaker_config_injection(sagemaker_sessio
             "KmsKeyId": expected_data_capture_kms_key_id,
         },
         KmsKeyId=expected_kms_key_id,
+        Tags=expected_tags,
+    )
+
+
+def test_create_endpoint_config_with_sagemaker_config_injection_no_kms_support(sagemaker_session):
+    """Tests that when no production variant instance supports KMS, no KMS key is used."""
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_ENDPOINT_CONFIG
+
+    data_capture_config_dict = {"DestinationS3Uri": "s3://test"}
+
+    # This method does not support ASYNC_INFERENCE_CONFIG or multiple PRODUCTION_VARIANTS
+    sagemaker_session.create_endpoint_config(
+        "endpoint-test",
+        "simple-model",
+        1,
+        "ml.g5.xlarge",
+        data_capture_config_dict=data_capture_config_dict,
+    )
+    expected_data_capture_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["DataCaptureConfig"]["KmsKeyId"]
+
+    expected_tags = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"]["Tags"]
+
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_with(
+        EndpointConfigName="endpoint-test",
+        ProductionVariants=[
+            {
+                "ModelName": "simple-model",
+                "VariantName": "AllTraffic",
+                "InitialVariantWeight": 1,
+                "InitialInstanceCount": 1,
+                "InstanceType": "ml.g5.xlarge",
+            }
+        ],
+        DataCaptureConfig={
+            "DestinationS3Uri": "s3://test",
+            "KmsKeyId": expected_data_capture_kms_key_id,
+        },
         Tags=expected_tags,
     )
 
@@ -2736,6 +2874,129 @@ def test_create_endpoint_config_from_existing_with_sagemaker_config_injection(
     )
 
 
+def test_create_endpoint_config_from_existing_with_sagemaker_config_injection_partial_kms_support(
+    sagemaker_session,
+):
+    """Tests that when some production variant instances supports KMS, a KMS key is used.
+
+    Note that this will result in an Exception being thrown at runtime. However, this is a preferable
+    experience than not using KMS keys when one could be used, from a security perspective.
+    """
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_ENDPOINT_CONFIG
+
+    pvs = [
+        sagemaker.production_variant("A", "ml.g5.2xlarge"),
+        sagemaker.production_variant("B", "ml.p2.xlarge"),
+        sagemaker.production_variant("C", "ml.p2.xlarge"),
+    ]
+    # Add DestinationS3Uri to only one production variant
+    pvs[0]["CoreDumpConfig"] = {"DestinationS3Uri": "s3://test"}
+    existing_endpoint_arn = "arn:aws:sagemaker:us-west-2:123412341234:endpoint-config/foo"
+    existing_endpoint_name = "foo"
+    new_endpoint_name = "new-foo"
+    sagemaker_session.sagemaker_client.describe_endpoint_config.return_value = {
+        "ProductionVariants": [sagemaker.production_variant("A", "ml.m4.xlarge")],
+        "EndpointConfigArn": existing_endpoint_arn,
+        "AsyncInferenceConfig": {},
+    }
+    sagemaker_session.sagemaker_client.list_tags.return_value = {"Tags": []}
+
+    sagemaker_session.create_endpoint_config_from_existing(
+        existing_endpoint_name, new_endpoint_name, new_production_variants=pvs
+    )
+
+    expected_production_variant_0_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["ProductionVariants"][0]["CoreDumpConfig"]["KmsKeyId"]
+    expected_inference_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "AsyncInferenceConfig"
+    ]["OutputConfig"]["KmsKeyId"]
+    expected_volume_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "KmsKeyId"
+    ]
+    expected_tags = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"]["Tags"]
+
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_with(
+        EndpointConfigName=new_endpoint_name,
+        ProductionVariants=[
+            {
+                "CoreDumpConfig": {
+                    "KmsKeyId": expected_production_variant_0_kms_key_id,
+                    "DestinationS3Uri": pvs[0]["CoreDumpConfig"]["DestinationS3Uri"],
+                },
+                **sagemaker.production_variant("A", "ml.g5.2xlarge"),
+            },
+            {
+                # Merge shouldn't happen because input for this index doesn't have DestinationS3Uri
+                **sagemaker.production_variant("B", "ml.p2.xlarge"),
+            },
+            sagemaker.production_variant("C", "ml.p2.xlarge"),
+        ],
+        KmsKeyId=expected_volume_kms_key_id,  # from config
+        Tags=expected_tags,  # from config
+        AsyncInferenceConfig={"OutputConfig": {"KmsKeyId": expected_inference_kms_key_id}},
+    )
+
+
+def test_create_endpoint_config_from_existing_with_sagemaker_config_injection_no_kms_support(
+    sagemaker_session,
+):
+    """Tests that when no production variant instance supports KMS, no KMS key is used."""
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_ENDPOINT_CONFIG
+
+    pvs = [
+        sagemaker.production_variant("A", "ml.g5.2xlarge"),
+        sagemaker.production_variant("B", "ml.g5.xlarge"),
+        sagemaker.production_variant("C", "ml.g5.xlarge"),
+    ]
+    # Add DestinationS3Uri to only one production variant
+    pvs[0]["CoreDumpConfig"] = {"DestinationS3Uri": "s3://test"}
+    existing_endpoint_arn = "arn:aws:sagemaker:us-west-2:123412341234:endpoint-config/foo"
+    existing_endpoint_name = "foo"
+    new_endpoint_name = "new-foo"
+    sagemaker_session.sagemaker_client.describe_endpoint_config.return_value = {
+        "ProductionVariants": [sagemaker.production_variant("A", "ml.m4.xlarge")],
+        "EndpointConfigArn": existing_endpoint_arn,
+        "AsyncInferenceConfig": {},
+    }
+    sagemaker_session.sagemaker_client.list_tags.return_value = {"Tags": []}
+
+    sagemaker_session.create_endpoint_config_from_existing(
+        existing_endpoint_name, new_endpoint_name, new_production_variants=pvs
+    )
+
+    expected_production_variant_0_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["ProductionVariants"][0]["CoreDumpConfig"]["KmsKeyId"]
+    expected_inference_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "AsyncInferenceConfig"
+    ]["OutputConfig"]["KmsKeyId"]
+
+    expected_tags = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"]["Tags"]
+
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_with(
+        EndpointConfigName=new_endpoint_name,
+        ProductionVariants=[
+            {
+                "CoreDumpConfig": {
+                    "KmsKeyId": expected_production_variant_0_kms_key_id,
+                    "DestinationS3Uri": pvs[0]["CoreDumpConfig"]["DestinationS3Uri"],
+                },
+                **sagemaker.production_variant("A", "ml.g5.2xlarge"),
+            },
+            {
+                # Merge shouldn't happen because input for this index doesn't have DestinationS3Uri
+                **sagemaker.production_variant("B", "ml.g5.xlarge"),
+            },
+            sagemaker.production_variant("C", "ml.g5.xlarge"),
+        ],
+        Tags=expected_tags,  # from config
+        AsyncInferenceConfig={"OutputConfig": {"KmsKeyId": expected_inference_kms_key_id}},
+    )
+
+
 def test_endpoint_from_production_variants_with_sagemaker_config_injection(
     sagemaker_session,
 ):
@@ -2788,6 +3049,135 @@ def test_endpoint_from_production_variants_with_sagemaker_config_injection(
         ProductionVariants=expected_pvs,
         Tags=expected_tags,  # from config
         KmsKeyId=expected_kms_key_id,  # from config
+        AsyncInferenceConfig=expected_async_inference_config_dict,
+        DataCaptureConfig={"KmsKeyId": expected_data_capture_kms_key_id},
+    )
+    sagemaker_session.sagemaker_client.create_endpoint.assert_called_with(
+        EndpointConfigName="some-endpoint",
+        EndpointName="some-endpoint",
+        Tags=expected_tags,  # from config
+    )
+
+
+def test_endpoint_from_production_variants_with_sagemaker_config_injection_partial_kms_support(
+    sagemaker_session,
+):
+    """Tests that when some production variants support KMS, a KMS key is applied from config.
+
+    Note that this will result in an Exception being thrown at runtime. However, this is a preferable
+    experience than not using KMS keys when one could be used, from a security perspective.
+    """
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_ENDPOINT_CONFIG
+
+    sagemaker_session.sagemaker_client.describe_endpoint = Mock(
+        return_value={"EndpointStatus": "InService"}
+    )
+    pvs = [
+        sagemaker.production_variant("A", "ml.g5.xlarge"),
+        sagemaker.production_variant("B", "ml.p2.xlarge"),
+        sagemaker.production_variant("C", "ml.p2.xlarge"),
+    ]
+    # Add DestinationS3Uri to only one production variant
+    pvs[0]["CoreDumpConfig"] = {"DestinationS3Uri": "s3://test"}
+    sagemaker_session.endpoint_from_production_variants(
+        "some-endpoint",
+        pvs,
+        data_capture_config_dict={},
+        async_inference_config_dict=AsyncInferenceConfig()._to_request_dict(),
+    )
+    expected_data_capture_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["DataCaptureConfig"]["KmsKeyId"]
+    expected_inference_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "AsyncInferenceConfig"
+    ]["OutputConfig"]["KmsKeyId"]
+    expected_volume_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "KmsKeyId"
+    ]
+    expected_tags = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"]["Tags"]
+
+    expected_async_inference_config_dict = AsyncInferenceConfig()._to_request_dict()
+    expected_async_inference_config_dict["OutputConfig"]["KmsKeyId"] = expected_inference_kms_key_id
+    expected_pvs = [
+        sagemaker.production_variant("A", "ml.g5.xlarge"),
+        sagemaker.production_variant("B", "ml.p2.xlarge"),
+        sagemaker.production_variant("C", "ml.p2.xlarge"),
+    ]
+    # Add DestinationS3Uri, KmsKeyId to only one production variant
+    expected_production_variant_0_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["ProductionVariants"][0]["CoreDumpConfig"]["KmsKeyId"]
+    expected_pvs[0]["CoreDumpConfig"] = {
+        "DestinationS3Uri": "s3://test",
+        "KmsKeyId": expected_production_variant_0_kms_key_id,
+    }
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_with(
+        EndpointConfigName="some-endpoint",
+        ProductionVariants=expected_pvs,
+        Tags=expected_tags,  # from config
+        KmsKeyId=expected_volume_kms_key_id,  # from config
+        AsyncInferenceConfig=expected_async_inference_config_dict,
+        DataCaptureConfig={"KmsKeyId": expected_data_capture_kms_key_id},
+    )
+    sagemaker_session.sagemaker_client.create_endpoint.assert_called_with(
+        EndpointConfigName="some-endpoint",
+        EndpointName="some-endpoint",
+        Tags=expected_tags,  # from config
+    )
+
+
+def test_endpoint_from_production_variants_with_sagemaker_config_injection_no_kms_support(
+    sagemaker_session,
+):
+    """Tests that when no production variants support KMS, no KMS key is applied from config."""
+
+    sagemaker_session.sagemaker_config = SAGEMAKER_CONFIG_ENDPOINT_CONFIG
+
+    sagemaker_session.sagemaker_client.describe_endpoint = Mock(
+        return_value={"EndpointStatus": "InService"}
+    )
+    pvs = [
+        sagemaker.production_variant("A", "ml.g5.xlarge"),
+        sagemaker.production_variant("B", "ml.g5.xlarge"),
+        sagemaker.production_variant("C", "ml.g5.xlarge"),
+    ]
+    # Add DestinationS3Uri to only one production variant
+    pvs[0]["CoreDumpConfig"] = {"DestinationS3Uri": "s3://test"}
+    sagemaker_session.endpoint_from_production_variants(
+        "some-endpoint",
+        pvs,
+        data_capture_config_dict={},
+        async_inference_config_dict=AsyncInferenceConfig()._to_request_dict(),
+    )
+    expected_data_capture_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["DataCaptureConfig"]["KmsKeyId"]
+    expected_inference_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"][
+        "AsyncInferenceConfig"
+    ]["OutputConfig"]["KmsKeyId"]
+
+    expected_tags = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"]["EndpointConfig"]["Tags"]
+
+    expected_async_inference_config_dict = AsyncInferenceConfig()._to_request_dict()
+    expected_async_inference_config_dict["OutputConfig"]["KmsKeyId"] = expected_inference_kms_key_id
+    expected_pvs = [
+        sagemaker.production_variant("A", "ml.g5.xlarge"),
+        sagemaker.production_variant("B", "ml.g5.xlarge"),
+        sagemaker.production_variant("C", "ml.g5.xlarge"),
+    ]
+    # Add DestinationS3Uri, KmsKeyId to only one production variant
+    expected_production_variant_0_kms_key_id = SAGEMAKER_CONFIG_ENDPOINT_CONFIG["SageMaker"][
+        "EndpointConfig"
+    ]["ProductionVariants"][0]["CoreDumpConfig"]["KmsKeyId"]
+    expected_pvs[0]["CoreDumpConfig"] = {
+        "DestinationS3Uri": "s3://test",
+        "KmsKeyId": expected_production_variant_0_kms_key_id,
+    }
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_with(
+        EndpointConfigName="some-endpoint",
+        ProductionVariants=expected_pvs,
+        Tags=expected_tags,  # from config
         AsyncInferenceConfig=expected_async_inference_config_dict,
         DataCaptureConfig={"KmsKeyId": expected_data_capture_kms_key_id},
     )
@@ -4314,6 +4704,7 @@ def create_inference_recommendations_job_default_happy_response():
             "ModelPackageVersionArn": IR_MODEL_PACKAGE_VERSION_ARN,
         },
         "JobDescription": "#python-sdk-create",
+        "Tags": [{"Key": "ClientType", "Value": "PythonSDK-RightSize"}],
     }
 
 
@@ -4338,6 +4729,7 @@ def create_inference_recommendations_job_default_model_name_happy_response():
             "ModelName": IR_MODEL_NAME,
         },
         "JobDescription": "#python-sdk-create",
+        "Tags": [{"Key": "ClientType", "Value": "PythonSDK-RightSize"}],
     }